[machO] Set section to max alignment of content decls

+++ src/elf.rs
@@ -202,7 +202,7 @@ struct SectionBuilder {
     alloc: bool,
     size: u64,
     name_offset: usize,
-    align: Option<usize>,
+    align: Option<u64>,
 }
 
 impl SectionBuilder {
@@ -234,7 +234,7 @@ impl SectionBuilder {
         self
     }
     /// Specify section alignment
-    pub fn align(mut self, align: Option<usize>) -> Self {
+    pub fn align(mut self, align: Option<u64>) -> Self {
         self.align = align;
         self
     }
